+++ src/tkc/value.c
@@ -993,3 +993,220 @@ value_t* value_set_bitmap(value_t* v, void* bitmap) {
 
   return value_init(v, VALUE_TYPE_BITMAP);
 }
+/*operations*/ 
+
+ret_t value_lshift(value_t* v, value_t* result, uint32_t n) {
+  ret_t ret = RET_OK;
+  return_value_if_fail(v != NULL && result != NULL, RET_BAD_PARAMS);
+
+  switch (v->type) {
+    case VALUE_TYPE_INT8: {
+      value_set_int8(result, value_int8(v) << n);
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      value_set_uint8(result, value_uint8(v) << n);
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      value_set_int16(result, value_int16(v) << n);
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      value_set_uint16(result, value_uint16(v) << n);
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      value_set_int32(result, value_int32(v) << n);
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      value_set_uint32(result, value_uint32(v) << n);
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      value_set_int64(result, value_int64(v) << n);
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      value_set_uint64(result, value_uint64(v) << n);
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}
+
+ret_t value_rshift(value_t* v, value_t* result, uint32_t n) {
+  ret_t ret = RET_OK;
+  return_value_if_fail(v != NULL && result != NULL, RET_BAD_PARAMS);
+
+  switch (v->type) {
+    case VALUE_TYPE_INT8: {
+      value_set_int8(result, value_int8(v) >> n);
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      value_set_uint8(result, value_uint8(v) >> n);
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      value_set_int16(result, value_int16(v) >> n);
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      value_set_uint16(result, value_uint16(v) >> n);
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      value_set_int32(result, value_int32(v) >> n);
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      value_set_uint32(result, value_uint32(v) >> n);
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      value_set_int64(result, value_int64(v) >> n);
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      value_set_uint64(result, value_uint64(v) >> n);
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}
+
+ret_t value_lshift_r(value_t* v, value_t* result, uint32_t n) {
+  ret_t ret = RET_OK;
+  return_value_if_fail(v != NULL && result != NULL, RET_BAD_PARAMS);
+
+  switch (v->type) {
+    case VALUE_TYPE_INT8: {
+      int8_t vv = value_int8(v);
+      n = n % 8;
+      value_set_int8(result, (vv >> (8 - n)) | (vv << n));
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      uint8_t vv = value_uint8(v);
+      n = n % 8;
+      value_set_uint8(result, (vv >> (8 - n)) | (vv << n));
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      int16_t vv = value_int16(v);
+      n = n % 16;
+      value_set_int16(result, (vv >> (16 - n)) | (vv << n));
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      uint16_t vv = value_uint16(v);
+      n = n % 16;
+      value_set_uint16(result, (vv >> (16 - n)) | (vv << n));
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      int32_t vv = value_int32(v);
+      n = n % 32;
+      value_set_int32(result, (vv >> (32 - n)) | (vv << n));
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      uint32_t vv = value_uint32(v);
+      n = n % 32;
+      value_set_uint32(result, (vv >> (32 - n)) | (vv << n));
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      int64_t vv = value_int64(v);
+      n = n % 64;
+      value_set_int64(result, (vv >> (64 - n)) | (vv << n));
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      uint64_t vv = value_uint64(v);
+      n = n % 64;
+      value_set_uint64(result, (vv >> (64 - n)) | (vv << n));
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}
+
+ret_t value_rshift_r(value_t* v, value_t* result, uint32_t n) {
+  ret_t ret = RET_OK;
+  return_value_if_fail(v != NULL && result != NULL, RET_BAD_PARAMS);
+
+  switch (v->type) {
+    case VALUE_TYPE_INT8: {
+      int8_t vv = value_int8(v);
+      n = n % 8;
+      value_set_int8(result, (vv << (8 - n)) | (vv >> n));
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      uint8_t vv = value_uint8(v);
+      n = n % 8;
+      value_set_uint8(result, (vv << (8 - n)) | (vv >> n));
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      int16_t vv = value_int16(v);
+      n = n % 16;
+      value_set_int16(result, (vv << (16 - n)) | (vv >> n));
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      uint16_t vv = value_uint16(v);
+      n = n % 16;
+      value_set_uint16(result, (vv << (16 - n)) | (vv >> n));
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      int32_t vv = value_int32(v);
+      n = n % 32;
+      value_set_int32(result, (vv << (32 - n)) | (vv >> n));
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      uint32_t vv = value_uint32(v);
+      n = n % 32;
+      value_set_uint32(result, (vv << (32 - n)) | (vv >> n));
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      int64_t vv = value_int64(v);
+      n = n % 64;
+      value_set_int64(result, (vv << (64 - n)) | (vv >> n));
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      uint64_t vv = value_uint64(v);
+      n = n % 64;
+      value_set_uint64(result, (vv << (64 - n)) | (vv >> n));
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}